+++ branches/nhd2-exp/src/driver/rcinput.h
@@ -145,8 +145,8 @@
 
 
 
-typedef uint32_t neutrino_msg_t;
-typedef uint32_t neutrino_msg_data_t;
+typedef size_t neutrino_msg_t;
+typedef size_t neutrino_msg_data_t;
 
 #define NEUTRINO_UDS_NAME 		"/tmp/neutrino.sock"
 #define NEUTRINO_KEYMAP_FILE		CONFIGDIR "/keymap.conf"
